requirements.txt
clearml==1.9.1
dask==2023.2.1
hydra-core==1.3.2
matplotlib==3.7.0
moviepy==1.0.3
numba==0.56.4
numpy==1.23.5
pandas==1.5.3
pygame==2.2.0
scipy==1.10.1
seaborn==0.12.2
Shapely==2.0.1
tqdm==4.64.1
